// app/store/user/user.slice.ts
import { getStoreLocal } from '@/utils/local-storage';
import { createSlice } from '@reduxjs/toolkit';
import { checkAuth, loggout, login, register } from './user.actions';
import { IInitialState } from './user.interface';

const initialState: IInitialState = {
  user: getStoreLocal('user'),
  isLoading: false,
};

export const userSlice = createSlice({
  name: 'user',
  initialState,
  reducers: {},
  extraReducers: (builder) => {
    builder
      .addCase(register.pending, (state) => {
        state.isLoading = true;
      })
      .addCase(register.fulfilled, (state, action) => {
        state.isLoading = false;
        state.user = action.payload.user;
      })
      .addCase(register.rejected, (state) => {
        state.isLoading = false;
      })
      .addCase(login.pending, (state) => {
        state.isLoading = true;
      })
      .addCase(login.fulfilled, (state, action) => {
        state.isLoading = false;
        state.user = action.payload.user;
      })
      .addCase(login.rejected, (state) => {
        state.isLoading = false;
      })
      .addCase(loggout.fulfilled, (state) => {
        state.isLoading = false;
        state.user = null;
      })
      .addCase(checkAuth.fulfilled, (state, { payload }) => {
        state.isLoading = false;
        if (payload) {
          state.user = payload.user;
        }
      });
  },
});
